feature: create profile screen and bottom nav update
