+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,4 @@
 <resources>
     <string name="app_name">FinSmart</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="welcome">Welcome,</string>
 </resources>
